Deleted unused variables and added explicit type casts in order to remove warnings. There are two global variables left that cannot be declared as constants: last_texY and odd_table in OpenGL.h. Could it be possible to move them where they are used and made them local variables?

+++ Graphics/OpenGLFunctorI.h
@@ -287,7 +287,7 @@ TreeCompartment<TS,BUD>* DrawNeedlesFunctor<TS,BUD>::operator()(TreeCompartment<
 	  
 	  glDisable(GL_LIGHTING);
 	  draw_texfoliage_planes(length, radius, GetValue(*cfts, LGAWf), 
-				 GetValue(*cfts, LGAage));
+				 static_cast<int>(GetValue(*cfts, LGAage)));
 	  glEnable(GL_LIGHTING);
 	  glPopMatrix();
 	}
